move read_tfrecords to brain_generator to keep writing/reading in one place
